I happened to start creating base classes for the collision detection because they are necessary for having tanks and towers in the same base class. -------
I created the base class projectile that should be implemented by every
projectile, namely the normal bullet and the missile(at the moment we
don't have any other projectiles). And I created the base class target
that should be implemented by every entity that can be targeted such as
the tower and the small tank (probably there are other tanks to come).

I had to create those classes because I want to make shooting and moving
the targets more general and the handling of the AI's can be made more
generic. Furthermore the control of the tank could be probably extended
to any other target.